rename: 필요없는 파일 삭제 및 패키지 이동
