+++ app/src/main/java/woowacourse/shopping/di/Container.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.common
+package woowacourse.shopping.di
 
 import kotlin.reflect.KClass
