+++ app/src/main/java/woowacourse/shopping/di/ViewModelInjector.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import woowacourse.shopping.ui.common.Container
 import kotlin.reflect.full.primaryConstructor
 import kotlin.reflect.jvm.jvmErasure
