+++ app/src/main/java/woowacourse/shopping/ui/ShoppingApplication.kt
@@ -5,7 +5,7 @@ import woowacourse.shopping.data.CartRepository
 import woowacourse.shopping.data.CartRepositoryImpl
 import woowacourse.shopping.data.ProductRepository
 import woowacourse.shopping.data.ProductRepositoryImpl
-import woowacourse.shopping.ui.common.Container
+import woowacourse.shopping.di.Container
 
 class ShoppingApplication : Application() {
     override fun onCreate() {
